+++ packages/user/src/index.ts
@@ -73,6 +73,13 @@ declare module "@dzangolab/fastify-config" {
         name?: string;
       };
       features?: {
+        profileValidation?: {
+          /**
+           * @default false
+           */
+          enabled?: boolean;
+          fields?: Array<keyof User>;
+        };
         signUp?: {
           /**
            * @default true
@@ -119,6 +126,7 @@ export { default as areRolesExist } from "./supertokens/utils/areRolesExist";
 export { default as validateEmail } from "./validator/email";
 export { default as validatePassword } from "./validator/password";
 export { default as hasUserPermission } from "./lib/hasUserPermission";
+export { default as ProfileValidationClaim } from "./supertokens/utils/profileValidationClaim";
 export { default as userSchema } from "./graphql/schema";
 
 export * from "./constants";